+++ MarioProject/BrickAniIDs.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "AnimationIDs.h"
 
-constexpr int ID_ANI_BREAK_BRICK_EFFECT = 1e6 + 2e4 + 1e3;
-constexpr int ID_ANI_DESTROYABLE_BRICK = 1e6 + 2e4;
+constexpr int ID_ANI_BREAK_BRICK_EFFECT = static_cast<int>(1e6 + 2e4 + 1e3);
+constexpr int ID_ANI_DESTROYABLE_BRICK = static_cast<int>(1e6 + 2e4);
 
-constexpr int ID_ANI_ITEM_BRICK = 2e6 + 2e4;
-constexpr int ID_ANI_EMPTY_ITEM_BRICK = 2e6 + 2e4 + 5;
+constexpr int ID_ANI_ITEM_BRICK = static_cast<int>(2e6 + 2e4);
+constexpr int ID_ANI_EMPTY_ITEM_BRICK = static_cast<int>(2e6 + 2e4 + 5);
