+++ MarioProject/GoombaAniIDs.h
@@ -7,7 +7,7 @@ constexpr int ID_ANI_GOOMBA_WALKING = (ID_ANI_GOOMBA + 1000);
 constexpr int ID_ANI_GOOMBA_DIE = (ID_ANI_GOOMBA + 1001);
 
 
-constexpr int ID_ANI_RED_GOOMBA				= 1e6 + 5e4;
-constexpr int ID_ANI_RED_GOOMBA_WALKING		= 1e6 + 5e4;
-constexpr int ID_ANI_RED_GOOMBA_DIE			= 1e6 + 5e4 + 10;
-constexpr int ID_ANI_RED_PARA_GOOMBA		= 1e6 + 5e4 + 1e2;
+constexpr int ID_ANI_RED_GOOMBA				= static_cast<int>(1e6 + 5e4);
+constexpr int ID_ANI_RED_GOOMBA_WALKING		= static_cast<int>(1e6 + 5e4);
+constexpr int ID_ANI_RED_GOOMBA_DIE			= static_cast<int>(1e6 + 5e4 + 10);
+constexpr int ID_ANI_RED_PARA_GOOMBA		= static_cast<int>(1e6 + 5e4 + 1e2);
